Refactor e2e test files
Splits the e2e_test file into separate test files to make it easier to
add additional tests.
